Fixes a circular import and adds clib loading to fx_utils.py.
