+++ Python/flexsea/flexsea/fx_plotting.py
@@ -33,13 +33,19 @@ def plot_setpoint_vs_desired(
 
     # Current controller
     if ctrl_type == fxe.HSS_CURRENT:
-        title_str = "Current control with {:.1f} Hz, {:.1f} mA {} and {:.0f} Hz commands (ID {})"
+        title_str = (
+            "Current control with {:.1f} Hz, {:.1f} mA {} and {:.0f} Hz "
+            "commands (ID {})"
+        )
         title = title_str.format(sig_freq, sig_amplitude, sig_type, cmd_freq, dev_id)
         plt.ylabel("Motor current (mA)")
 
     # Position controller
     elif ctrl_type == fxe.HSS_POSITION:
-        title_str = "Position control with {:.1f} Hz, {:.0f} tick{} {} and {:.0f} Hz commands (ID {})"
+        title_str = (
+            "Position control with {:.1f} Hz, {:.0f} tick{} {} and {:.0f} Hz "
+            "commands (ID {})"
+        )
         title = title_str.format(
             sig_freq,
             sig_amplitude,
