+++ Python/flexsea/flexsea/fx_utils.py
@@ -1,6 +1,7 @@
 """
 General purpose utilities
 """
+import ctypes as c
 import os
 import platform
 
@@ -383,7 +384,7 @@ def print_net_master(net_master_state: fx_devs.NetMasterState):
 def print_bms(dev_id):
     """Print BMS info"""
     # TODO (CA): Implement this function
-    print("Printing BMS information not implemented. Device {}".format(dev_id))
+    print(f"Printing BMS information not implemented. Device {dev_id}")
 
 
 # ============================================
@@ -393,7 +394,7 @@ def print_loop_count(count, total):
     """
     Convenience function for printing run counts
     """
-    print("\nRun {} of {}".format(count + 1, total))
+    print(f"\nRun {count + 1} of {total}")
 
 
 # ============================================
@@ -403,11 +404,7 @@ def print_loop_count_and_time(count, total, elapsed_time):
     """
     Convenience function for printing run counts and elapsed time in s.
     """
-    print(
-        "\nLoop {} of {} - Elapsed time: {}s".format(
-            count + 1, total, round(elapsed_time)
-        )
-    )
+    print(f"\nLoop {count + 1} of {total} - Elapsed time: {round(elapsed_time)}s")
 
 
 # ============================================
@@ -418,7 +415,7 @@ def sin_generator(amplitude, freq, command_freq):
     Generate a sine wave of a specific amplitude and frequency
     """
     num_samples = command_freq / freq
-    print("number of samples is: ", int(num_samples))
+    print(f"number of samples is: {int(num_samples)}")
     in_array = np.linspace(-np.pi, np.pi, int(num_samples))
     sin_vals = amplitude * np.sin(in_array)
     return sin_vals
@@ -459,7 +456,7 @@ def load_ports_from_file(file_name):
             return vals["ports"], int(vals["baud_rate"])
 
     except IOError as err:
-        print("Problem loading {}: {}".format(file_name, err))
+        print(f"Problem loading {file_name}: {err}")
         print(
             "Copy the ports_template.yaml to a file named ports.yaml"
             "Be sure to use the same format of baud rate on the first line,"
@@ -467,5 +464,212 @@ def load_ports_from_file(file_name):
         )
         raise err
     except ValueError as err:
-        print("Problem with the yaml file syntax or values: {}", err)
+        print(f"Problem with the yaml file syntax or values: {err}")
         raise err
+
+
+# ============================================
+#                  _get_os
+# ============================================
+def _get_os() -> str:
+    """
+    Determines which operating system we're running on.
+
+    Returns
+    -------
+    str:
+            The name of the operating system. Matches the directory inside
+            `libs` containing the c libraries appropriate for that OS.
+    """
+    operating_system = "linux"
+    # Windows
+    if "win" in platform.system().lower():
+        # 32 bit
+        if platform.architecture()[0] == "32bit":
+            operating_system = "win32"
+        # 64 bit
+        else:
+            operating_system = "win64"
+    # Pi - 32 bit
+    elif os.uname().machine.startswith("arm"):
+        operating_system = "raspberryPi"
+    # Pi - 64 bit
+    elif os.uname().machine.startswith("aarch64"):
+        operating_system = "raspberryPi64"
+    else:
+        raise OSError("Unknown operating system!")
+    return operating_system
+
+
+# ============================================
+#                  _add_dlls
+# ============================================
+def _add_dlls(lib_path: str) -> None:
+    """
+    Explicitly adds the appropriate directories to the `PATH` on Windows.
+
+    Parameters
+    ----------
+    lib_path : str
+            The full path to the directory containing the Windows
+            precompiled libraries.
+    """
+    for extra_path in os.environ["PATH"].split(";"):
+        if os.path.exists(extra_path) and "mingw" in extra_path:
+            os.add_dll_directory(extra_path)
+    os.add_dll_directory(lib_path)
+
+
+# ============================================
+#             _set_data_types
+# ============================================
+def _set_data_types(clib: c.CDLL) -> c.CDLL:
+    """
+    Defines the function prototypes.
+
+    Parameters
+    ----------
+    clib : c.CDLL
+            The Python object interfacing with the C libraries we're
+            setting the prototypes for.
+
+    Returns
+    -------
+    clib : c.CDLL
+            The Python object interfacing with the C libraries for which
+            we've set the prototypes.
+    """
+    clib.fxOpen.argtypes = [c.c_char_p, c.c_uint, c.c_uint]
+    clib.fxOpen.restype = c.c_int
+
+    clib.fxIsOpen.argtypes = [c.c_uint]
+    clib.fxIsOpen.restype = c.c_bool
+
+    clib.fxClose.argtypes = [c.c_uint]
+    clib.fxClose.restype = c.c_int
+
+    clib.fxCloseAll.argtypes = []
+    clib.fxCloseAll.resType = []
+
+    clib.fxGetDeviceIds.argtypes = [c.POINTER(c.c_int), c.c_uint]
+
+    clib.fxStartStreaming.argtypes = [c.c_uint, c.c_uint, c.c_bool]
+    clib.fxStartStreaming.restype = c.c_int
+
+    clib.fxStopStreaming.argtypes = [c.c_uint]
+    clib.fxStopStreaming.restype = c.c_int
+
+    clib.fxReadDevice.argtypes = [c.c_uint, c.POINTER(fx_devs.ActPackState)]
+    clib.fxReadDevice.restype = c.c_int
+
+    clib.fxReadDeviceAll.argtypes = [
+        c.c_uint,
+        c.POINTER(fx_devs.ActPackState),
+        c.c_uint,
+    ]
+    clib.fxReadDeviceAll.restype = c.c_int
+
+    clib.fxReadMdDeviceAll.argtypes = [
+        c.c_uint,
+        c.POINTER(fx_devs.MD10State),
+        c.c_uint,
+    ]
+    clib.fxReadMdDeviceAll.restype = c.c_int
+    clib.fxReadMdDevice.argtypes = [c.c_uint, c.POINTER(fx_devs.MD10State)]
+    clib.fxReadMdDeviceAll.restype = c.c_int
+
+    clib.fxReadNetMasterDevice.argtypes = [c.c_uint, c.POINTER(fx_devs.NetMasterState)]
+    clib.fxReadDevice.restype = c.c_int
+
+    clib.fxReadNetMasterDeviceAll.argtypes = [
+        c.c_uint,
+        c.POINTER(fx_devs.NetMasterState),
+        c.c_uint,
+    ]
+    clib.fxReadNetMasterDeviceAll.restype = c.c_int
+
+    clib.fxSetReadDataQueueSize.argtypes = [c.c_uint, c.c_uint]
+    clib.fxSetReadDataQueueSize.restype = c.c_uint
+
+    clib.fxGetReadDataQueueSize.argtypes = [c.c_uint]
+    clib.fxGetReadDataQueueSize.restype = c.c_int
+
+    clib.fxSetGains.argtypes = [
+        c.c_uint,
+        c.c_uint,
+        c.c_uint,
+        c.c_uint,
+        c.c_uint,
+        c.c_uint,
+        c.c_uint,
+    ]
+    clib.fxSetGains.restype = c.c_int
+
+    clib.fxSendMotorCommand.argtypes = [c.c_uint, c.c_int, c.c_int]
+    clib.fxSendMotorCommand.restype = c.c_int
+
+    clib.fxGetAppType.argtypes = [c.c_uint]
+
+    clib.fxActivateBootloader.argtypes = [c.c_uint, c.c_uint8]
+    clib.fxActivateBootloader.restype = c.c_int
+
+    clib.fxIsBootloaderActivated.argtypes = [c.c_uint]
+    clib.fxIsBootloaderActivated.restype = c.c_int
+
+    clib.fxRequestFirmwareVersion.argtypes = [c.c_uint]
+    clib.fxRequestFirmwareVersion.restype = c.c_int
+
+    clib.fxGetLastReceivedFirmwareVersion.argtypes = [c.c_uint]
+    clib.fxGetLastReceivedFirmwareVersion.restype = fxe.FW
+
+    return clib
+
+
+# ============================================
+#                 _load_clib
+# ============================================
+def _load_clib() -> c.CDLL:
+    """
+    Uses `ctypes` to load the appropriate C libraries depending on the
+    OS.
+
+    Raises
+    ------
+    OSError:
+            Raised if the precompiled libraries fail to load.
+
+    Returns
+    -------
+    clib : ctypes.cdll.CDLL
+            The Python object from which we can call the flexsea C
+            functions.
+    """
+    nix_lib = "libfx_plan_stack.so"
+    win_lib = "libfx_plan_stack.dll"
+    libs_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "libs")
+    os.environ["PATH"] += libs_path
+
+    operating_system = _get_os()
+    lib_path = os.path.join(libs_path, operating_system)
+
+    if "win" in operating_system:
+        try:
+            _add_dlls(lib_path)
+            clib = c.cdll.LoadLibrary(os.path.join(lib_path, win_lib))
+        except OSError as err:
+            msg = f"\n[!] Error loading the {win_lib} precompiled libraries.\n"
+            msg += (
+                "The most likely cause is a mismatch between the Python, pip and shell "
+                "architectures.\n"
+            )
+            msg += "To solve this, ensure all three are either 32 or 64 bit.\n"
+            msg += "Keep different versions isolated by virtual environments.\n"
+            msg += "For additional help, follow the installation instructions here: "
+            msg += "https://github.com/DephyInc/Actuator-Package\n"
+            msg += "--------------------------------------------\n\n"
+            msg += f"Detailed error message for debugging:\n{err}\n"
+            print(msg)
+    else:
+        clib = c.cdll.LoadLibrary(os.path.join(lib_path, nix_lib))
+
+    return _set_data_types(clib)
